// typechain-abi/BalancerV2VaultAbi.d.ts
export interface BalancerV2VaultAbi extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): BalancerV2VaultAbi;
  clone(): BalancerV2VaultAbi;
  methods: {
// ...
    getPausedState(): NonPayableTransactionObject<{
      paused: boolean;
      pauseWindowEndTime: string;
      bufferPeriodEndTime: string;
      0: boolean;
      1: string;
      2: string;
    }>;
// ...
  };
// ...
}